[FIX][BE][석범] 완료된 사람이 시간 정보 req 추가

+++ backend/frazzle/src/main/java/com/frazzle/main/domain/socket/game/dto/EndResponseDto.java
@@ -10,16 +10,16 @@ import java.util.List;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class EndResponseDto {
-    private Long time;
+    private int time;
     private String nickname;
 
     @Builder
-    private EndResponseDto(Long time, String nickname) {
+    private EndResponseDto(int time, String nickname) {
         this.time = time;
         this.nickname = nickname;
     }
 
-    public static EndResponseDto createEndResponseDto(Long time, String nickname) {
+    public static EndResponseDto createEndResponseDto(int time, String nickname) {
         return EndResponseDto.builder()
                 .time(time)
                 .nickname(nickname)
